Add files via upload
A validation error occurs during rendering. The error states "... Descriptor in binding #0 index 1 is being used in draw but has never been updated via
vkUpdateDescriptorSets() or a similar call. ...", however, the application is running perfectly fine. I believe there is a error with the validation layer. This should not break anything and may be used (with caution as I do not know if any thing wacky is happening behind the scenes)